Move additional codegen helpers
